HUB-555: Added a /completed page for IDPs to return the users
If the IDP no longer has the original SAML request for the user, they're not able
to issue a SAML response back to the hub. This new page will allow them to send the
user back. Also the page updates the journey hint to hint the user next time round.