HUB-566: Change how AB test is reported
Piwik/Matomo tracks what variant the user is in during an AB test. This is achieved by setting a custom
variable for AB_TEST. Currently this gets set by reporting a virtual page view
"The user has started an AB test" at the start of the test.

This change stops using the extra call to piwik, but instead adds the custom variable
our existing array (along with LOA and RP). Which gets reported with a standard page view.
This eliminates the race condition between the virtual page view and a standard page view.

It's using flash on the request, as it's not needed to persist across and is only required
when user visits the routes which are being AB-tested.

Additionally, it also simplifies the set up of the AB test, since we no longer need
two set of constraints for piwik reporting and normal routes.