HUB-583: Handle not available IDPs on paused/resume pages
Currently, a resume page will show for IDPs who are no longer accepting
registrations. This change is to make a switch to check the IDP against the list of registration
IDPs, instead of the one for sign-in. Similar change is on the paused page.
The IDP's simple ID gets now checked with the registration list and
the behaviour is the same when user types in a non-existent IDP