HUB-580: Enable running feature tests separately
The feature test do not work when run separately using `bundle exec rspec spec/features/some_feature_spec.rb`
This is due to the fact we introduced multithreading for in our loading_cache.
When Capybara runs it uses a single thread causing a deadlock since the
thread gets lost. This change allows the test boostrap the application including
the cache and threads.

This can also be enabled in the test.rb config globally, however we only need
it for the feature tests, so adding it to the feature_helper.