BAU: Cookie monster is now longer needed
Our cookie monster was implemented in Sep 2018. At that time we changed
the structure/schema of the journey hint cookie. It's been cleaning users'
cookie from the old structure/schema. Given the 18 months cookie
expiry, it's now to safe to remove it as there should be no more users with the old
schema.