HUB-518 Handle cases where the entityID found in the resume cookie is no longer valid
HMRC changed their entityIDs which means that some users with the resume cookie are getting an error page because the entityId isn't registered with Config.

Now the users get the same experience as they would if the IDP was no longer found.

We log this as a warning. If we don't log anything then the linter gets upset about us supressing exceptions and we can't explicitely return nil because the linter gets upset about that instead.