HUB-247 Build new hub journey that comes after failure
Present different content to users depending on how many IDPs are available to them to choose.

When a user fails registration, the IDP they failed with is saved into an array session variable `idps_tried`.  This array is used to instantiage a Set in order to determine how many untried IDPs remain which in turn determines which content is displayed.

Modify Gemfile because the debug view in stopped rendering and the browser's 500 page would show instead.