HUB-667 Record missing journey type when user follows hint on start page
When a user followed the IDP sign in hint shown on the start page, that was reflected in the call to policy and the user's journey type was not recorded.  This PR fixes that and clarifies the use of journey_type somewhat.

Reorganise constants used for journey_type and add one to specify the user has followed a hint

Apply new constants throughout code

Pass journey_type when users follows sign in hint on start page

Add tests to exercise this user journey

Co-authored-by: Olakunle Jegede <olakunle.jegede@digital.cabinet-office.gov.uk>
Co-authored-by: John Watts <john.watts@digital.cabinet-office.gov.uk>