Updated to Ruby 2.6.6
A few version bumps in the Gemfile as well as the bump in ruby.  Some of
this work is to get ready for the eventual upgrade to Rails 6.0 as well
as keeping things up to date.

Also pegged the version of things in the stub/api/Gemfile as they were
now pull things for Rails 6 and as a result pre-commit was failing.

We should probably audit the version of things in the Gemfile.