BAU: Fix Ruby 2.6.6 expecting a newer version of system library
After upgrading Ruby to 2.6.6, the gem ffi got updated to a newer version that requires a version of the system library `/usr/lib/libffi.dylib` that's not available on MacOS Mojave. This commit reverts the gem to a previous version that works with the library available on our dev machines.

Ideally we'd be running everything in Docker and this wouldn't be an issue but as a temporary solution we can use an older version of the library to make this work on our machines.