Ensure Pwiki SameSite cookie behaviour
This commit updates the piwik JavaScript to the latest version found at
https://github.com/matomo-org/matomo/blob/4.x-dev/js/piwik.min.js and
also adds the code to set the SameSite cookie behaviour.

As the new Matomo Javascript changes the interactions slightly and uses
post requests rather than get requests this also updates some of the
routes to support making post requests to the matomo stub we have in our
code base.