HUB-783: Don't return failed IDPs from recommendations engine
This fixes a bug in the frontend. If a user failed to register with an
IDP and their AuthnResponse was returned with a "FAILED" state, they
would be directed via a landing page and a link back to the choose a
company page.

The logic on that page would retrieve a list of IDPs from the
recommendation engine, including the IDP they just failed with. If they
the selected that IDP again, it's entity ID would be compared to a
separate list of available IDPs - this time with any IDPs the user had
already failed with removed. This caused an error.

This change ensures the recommendation engine doesn't return an IDP the
user has failed with, and so it's not presented to them on the choose an
IDP page.