HUB-740: Handle paused sign in journeys
Experian will allow, in certain circumstances, a user to pause their
sign in journey. This appears to be when they need to re-confirm their
evidence.

This change ensure that the correct page is shown to the user if they've
come from a sign in journey, rather than being shown a page about the
IDP no longer providing registrations.

Experian are disconnecting completely in 6 weeks so this seems like the
most productive approach, rather than asking them to fix upstream.