HUB-242: Make short-Hub page title more accessible
It is recommended that page titles and H1 should match
This requirement implies that page titles and H1 should be semantically
similar. Making page titles are exactly the same as there respective H1
would ensure that this accessibility criteria is always respected and
prevent any deviation and inconsistencies from setting in overtime.
The H1 is the definitive version, so all page titles are set to the
respective top level heading of the page.

This change does the following:
- all page titles are set to the content of there respective H1 heading
- all page title in `en.yml` and `cy.yml` have been removed
- the top level heading H1 names end with `.heading` this is to establish
  a convention