BAU: Remove unnecessary args to `translate`
We've bumped the version of the `i18n` module from 1.8.5 to 1.8.9.
Version [1.8.6](https://github.com/ruby-i18n/i18n/releases/tag/v1.8.6)
introduced a change to the parameters for the `translate` method. It
removed the `*` operator, which allows any number of args to be passed
in. These args were just discarded.

You can see the change to the source here: https://github.com/ruby-i18n/i18n/pull/545

We were passing in an extra arg in a couple of places in our tests. This
removes them and prevents an `ArgumentError` being thrown.