BAU: Only skip session validation where necessary
In the PausedRegistration controller we were skipping session
validation, except for the `resume` action. This was so that users
returning after pausing, via the #index or #from_resume_link actions,
who's sessions may have expired, weren't booted to the usual error page.

We were explicitly still validating the `resume` action however. This
action is accessed with a GET request. The `resume_with_idp` and
`resume_with_idp_ajax` actions are accessed with POST requests made from
the `resume-registration` page rendered by the `resume` action. We were
not validating the session on theses POSTs.

We've seen occurrences of a users session becoming invalid in between the
initial GET for `resume` and then the following POSTs. We should
validate the POSTs too to capture this.