BAU: Remove `RESUMELINK` from journey hint after CompletedRegistration
The `RESUMELINK` value is set in the journey hint controller when a
users registration is paused with a resume link. This value is checked
in the AuthnRequest controller and the user is redirected to the resume
registration path if it exists.

We unset the value once a user has completed a journey with an IDP -
this is done in the AuthnResponseController#idp_response action. This is
where IDP responses are POSTED to.

Some users however are returned by IDPs to the
CompletedRegistrationController. This is when the IDP no longer has the
original SAML request for the user, and so can't issue the response back
to the hub. This happens when a user resumes a paused registration
journey.

We weren't unsetting the RESUMELINK value in the journey hint cookie in
ths scenario. This was resulting in users being directed back down the
resume registration path when trying to sign in with a service, despite
having already registered.

This issue was noticed when investigating a bug with users not having
`transaction_entity_id` set in their session when hitting the
PausedRegistrationController#from_resume_link method. This is happening
because we're also setting `session[:transaction_entity_id] = nil` in
the CompletedRegistrationController (a hack). This change won't fix the
underlying issues, but it will stop us channeling users down that route.