HUB-971: Remove unused 'Why companies' page
It looks like this page was suppsed to be linked to from the IDP picker pages, as there were two duplicates for both LOAs. However, LOA2 IDP picker wasn't linking to this page and the LOA1 picker had broken text. The most reasonable course of action was to axe this page make LOA1 picker match LOA2 picker in content.