HUB-971: Introduce the new "about documents" page to replace the document picker page
The new page is just a static page that lets all users through without having to select any document they have. This means we no longer need the IDP recommendation engine, user segments and all the logic and config associated with these features. However, to keep the commits to a reasonable size, at this stage the new page is introduced without tearing down the exising components. This means we have to fake the selected documents answers for the time being.

  - Refactor the "about" controller
  - Add the "about documents" page to the about controller
  - Move the "prove your identity another way" page to the about controller
  - Redirect to the new page from the "will it work for me" page
  - Move all "about" content into one group in the language files

  - Fix a broken redirect for LOA1 journeys on the "about choosing a company" page

  - Remove some proxy node leftovers