BAU: Make tests in `user_visits_continue_to_your_idp_page_spec` work offline
The `will redirect the user to the IDP on Continue` test only works online as the redirect URL of the RP (www.gov.uk) must be reachable by Capybara/Selenium and the GET request must return a 200 response in order for the `click_button "initiate-single-idp-post"` action to succeed.

If the test is run offline, the redirect URL cannot be resolved and the test fails even though there are no bugs in the code.

To resolve this, set the redirect URL to the `/test-saml` page of the local Capybara server. This way, the URL can always be resolved regardless of the presence or lack of internet connection.