HBU-971 Cleanup: Remove the "select documents" page and the checkboxes form
The page is no longer used (replaced by static "about documents"), so we can remove the associated styles, views and logic.

  - Remove logic to report user evidence, as users no longer need to provide any evidence to get through to the IDP picker page

  - Render the "something went wrong page" if there no IDPs available for registration
Since we no longer filter IDPs based on the user's evidence, we will always show all available IDPs. If there are no IDPs, this is a critical state which would indicate there is a problem with hub federation config. This should never occur.

  - Use symbols and string literals in `piwik_test_helper` to improve readability