HUB-1117: One IDP updates to How Verify Works Page
Preparing for a one IDP world we've been asked to make some content
changes to the How Verify Works page.  This commit introduces an if
statement to the view which renders the existing content if there is
more than one page or the new content if only 1 IDP is present.